+++ src/db/index.js
@@ -1,0 +1,1 @@
+export { default as Repository } from './sequelize-repository'
